fix(cli): diff against multiple stacks should fail if any have a diff
Currently, the diff command only remembers if there are any diffs
with the last stack that was diffed. So, if all of your stacks have
a diff except for the last one, the diff command would exit
successfully even if told to fail on diff.

This fix corrects the behavior so that if any of the stacks have
a diff, then the exit code would be non-zero if configured to do so.

fixes #7492