chore(cli): install specific cli version in integ tests (#7704)
Currently, we install the CLI in the integration tests by simply running `npm install aws-cdk`.
We then make sure the version that gets installed is the version we have locally.

This will fail immediately after the release of a new version, and before the merge back PR was merged with the following error:

```console
|
| ============================================================================================
| Expected CDK version: 1.36.0
| ============================================================================================
| Found CDK: /tmp/cdk-rundist/node_modules/.bin/cdk
| Mismatched CDK version. Expected: 1.36.0, actual: 1.36.1
1.36.1 (build 4df7dac)
```

It happens because the verdaccio instance we have in the tests has an npm uplink configured, so the latest version as far as he is concerned is the latest published one, which doesn't match the version number we have locally since the merge back PR wasn't merged yet.

This PR makes it so we always install the CLI version we want to test (i.e the local version).