refactor(core): fold "assets" to "core"
Fold the "assets" module, which includes the Staging construct that takes care of staging asset files into the cloud assembly during synthesis into "core". This is in order to allow implementing custom resources that leverage assets throughout the framework.

A subsequent commit will add a mini-framework for custom resources that leverages this capability.