fix(cli): cdk deploy cannot update stacks in REVIEW_IN_PROGRESS status
Stacks only end up in `REVIEW_IN_PROGRESS` state when they have never been deployed.
It exists, but has never been deployed, and therefore has no resources or a template.

Users can end up in this state if a changeset takes too long to deploy or if the
`--no-execute` flag is used the first time a stack is deployed.

Root cause:
We're trying to `update` a stack that was never deployed.
CloudFormation considers a stack ID with no template or resources to be a stack that does not exist.
When we try to set the `ChangeSetType` to `UPDATE`, it results in a `ValidationException`

Fix:
When a stack is in `REVIEW_IN_PROGRESS` state, ensure we set the `ChangeSetType` to `CREATE`

Closes #6674