fix(ec2): can't add VPN connections to a VPC progressively
When creating a vpc that contains no subnets, you can not add a vpn gateway
or vpn connections to the vpc while creating. This is all good.
But I'm later adding subnets and want to add a vpn connection, then
it is not possible, the vpnGateay boolean in IVpc is a read only boolean
and is set to false and there's no way to change that. This commit adds
the method enableVpnGateway to later allow adding vpn connections when
startinh with a vpc with no subnets.