chore: move context queries to cloud assembly schema
There was another protocol of implicit "props" queries in
the cloud assembly, which is the "missing context queries"
protocol.

Move it into the cloud-assembly-schema package so we can
properly version updates to it.

There was still some additional stuff in the protocol around constants
and return types. Did not move those yet since I'm not sure
how they fit into the schema. Taking it one step at a time.