feat(core): custom resource provider helper
A helper for implementing simple node.js-based custom resource providers. This is a simpler framework from what is offered `@aws-cdk/custom-resources.Provider`, designed to enable implementing custom resources with minimal dependencies. To that end, this helper uses `CfnResource` to define the AWS Lambda function and the low-level asset mechanism in order to create an asset for the lambda bundle.

Comparing to the advanced provider framework, this provider DOES NOT support:
- Arbitrary lambda function handlers, only node.js function.
- Asynchronous "isComplete" waiters (limited to 15min lambda timeout).

Since `cdk-integ` depends on `cdk` which depends on `@aws-cdk/core` (as a "dev dependency"), I've added the integ test for this in @aws-cdk/aws-cloudformation. This probably belongs to some other module (`@aws-cdk/core-tests`?).

This is a precursor for implementing support for Open ID connect providers in the AWS IAM module, which is a very low-level module in our stack.