chore: force resolution of tap-mocha-reporter
We are inherited a transitive dependency on diff@1.4.0 which has been
flagged as having a security vulnerability (all versions <3.5.0 are
vulnerable).

This dependency is coming in via nodeunit → node-tap →
tap-mocha-reporter → diff.

Only the absolute latest release, 5.0.1 as of this commit, of
tap-mocha-reporter has upgraded to use diff@4.0.1.
Hence, forcing resolution to use tap-mocha-reporter@5.0.1 so we get rid
of the old version of diff.

In the longer term, nodeunit is deprecated and we should be moving to
jest which would prune out this dependency subtree.