fix(aws-batch): gpuCount was ignored in JobDefinition creation (#7587)
basically gpus were not being assigned to JobDefinitions because while the props accepted gpuCount, it was never passed down to the container properties and ultimately the resource requirements section of the Cfn resource.

This PR will fix this issue and actually allow Gpus to be set using the JobDefinition resource.