fix(eks): "vendor response doesn't contain attribute" when updating version (#7830)
Cluster version updates fail with `vendor response doesn't contain <ATTRIBUTE>` errors due to the fact that since #7526 the provider does not respond to `isComplete` with the `Data` field with resource attributes.

The fix is that once the update is complete, we simply delegate to `isActive` which queries the cluster and returns the attributes.

Fixes #7794