feat(codebuild): add support for test reports
This adds support for CodeBuild's test reports feature.
It includes a new Layer 2 class, ReportGroup.

The default project's IAM role has been augmented with permissions
to create and use test result groups whose names begin with the project's name
(as that is what the default report group created when putting a simple name in the
'reports' buildspec section is called).
There is a new boolean property when creating the project,
addCreateReportGroupPermissions,
that can be used to turn off this behavor.

Fixes #7367