feat(stepfunctions): custom state as an escape hatch
Custom State which enables the capability to provide Amazon States Language (ASL) JSON as an escape hatch. Useful when there are capabilities that are offered through Step Functions such as service integrations, and state properties but there isn't support through the CDK yet.

It enables the usage of all service integrations we don't currently support.