chore(cli): re-enable regression tests (#7829)
This PR re-enables the cli regression tests.

**Why were they disabled?**

We were booting up verdaccio without configuring an `npm` uplink for our own packages. This was ok because we used to only test locally built packages, both of the CLI and the framework.

The regression tests introduced another combination where we test the locally built CLI, against the latest published packages of the framework. But since we didn't used to have an uplink for our own packages. the installation would fail.

**Why will it work now?**

Two reasons:

1. We now do have an `npm` uplink for verdaccio.
2. Don't start a long running verdaccio instance for tests that use the latest published version.

In addition, this PR exposes various execution permutations via the `package.json` instead of hiding them in scripts. This brings it back to how it was originally.