chore: change CLI integ tests from Bash to TypeScript
Converted most tests from Bash to TypeScript. It's much more comfortable
to write complex tests in a programming language than bash. Plus, `jest`
is a better test runner than bash, too.

Even though tests are now written in TypeScript, this does not
conceptually change their SUT! They are still testing the CLI via
running it as a subprocess, they are NOT reaching directly into the CLI
code to test its private methods. It's just that setup/teardown/error
handling/assertions/AWS calls are much more convenient to do in a real
programming language.

Compilation of the tests is done as part of the normal package build, at
which point it is using the dependencies brought in by the containing
`aws-cdk` package's `package.json`.

When run in a non-develompent repo (as done during integ tests or canary
runs), the required dependencies are brought in just-in-time via
`test-jest.sh`. Using a separate `package.json` just for the tests
makes it a package in its own right for `yarn`, and I didn't want
to go there.