feat(lambda): docker code
Add `Code.fromDockerImage` and `Code.fromDockerAsset` that offer a generic way
of working with Docker for Lambda code.

`Code.fromDockerImage`: run a command in an existing Docker image
`Code.fromDockerAsset`: build an image and then run a command in it

Both `Code` classes take an `assetPath` prop that corresponds to the path of the
asset directory that will contain the build output of the Docker container. This
path is automatically mounted at `/asset` in the container. Using a combination
of image and command, the container is then responsible for putting content at
this location. Additional volumes can be mounted if needed.

This will allow to refactor `aws-lambda-nodejs` and create other language
specific modules.