chore: CDK apps generate asset manifest
Introducing new `StackSynthesizer` objects which can be attached to `Stacks`. StackSynthesizer
controls how a stack synthesizes itself to a Cloud Assembly, and specifically how it treats assets.

The most important Stack Synthesizers in this change are:

*  `LegacyStackSynthesizer`, the "old" stack synthesis behavior, synthesizing assets to parameters.
* `DefaultStackSynthesizer`, new behavior which goes hand-in-hand with bootstrapping v2, which synthesizes well-known locations and an asset manifest for assets.

The feature flag `@aws-cdk/core:newStyleStackSynthesis` controls whether an unconfigured stack gets the new or old synthesis object.

ALSO IN THIS PR

* Placeholder replacements for account, region and partition are used both by the CLI and the `cdk-assets` tool. Move the functionality for both into `cx-api` so as to not have to duplicate it.
* Artifact properties were an untyped property bag which should have gone into the Cloud Assembly Schema. Move them there.
* Split out the `schema.ts` file into smaller files to prevent it from getting too big.