chore(cfn2ts): support for filtering on resource name prefix
Add the ability to filter resource types based on its name, so that
L1 constructs that match up with the resource type name's prefix
are included in the generated typescript code.

Motivation
The CDK `aws-cognito` module contains two different product offerings -
user pool and identity pool. In order to track these effectively in the
CDK and configure the right level of support for each of these, we are
looking to split these into two packages `aws-cognito-userpool` and
`aws-cognito-identitypool`.