chore: fix integ tests
When introducing the Bash/TS tests, we added `*.js` to the `.gitignore`
(because the tests are written in TypeScript).

Because `.gitignore` patterns are implicitly included in `.npmignore`,
this led to the `manifest.json.js` files in the `cloud-assemblies`
directories to longer be included in the NPM tarballs, which led to the
integ tests failing.

Add repro instructions so I don't have to keep looking this up every
time, and adding a `--prefix` argument to npm install, which I needed
to add on my machine to make `run-against-dist` work.