chore: fix integ tests, take 3851
More `gitignore`/`npmignore` woes, still dropping more `.js` files from
the tarball than we intended. Consolidate all `gitignore` instructions
in a single file at the root of the package, so it's very clear which
files are going to be honored.

While working on this, figured out that the `run-against-dist` script
was trampling one EXIT trap with another (leading to NuGet config being
restored but the local NPM registry NOT being killed), so add a
mechanism so we can have arbitrarily many EXIT traps.