chore: update supported node versions (#7887)
This is a follow-up to the recent issue we had where `uuid` had bumped up versions and that included changes that affected earlier version of node `13.x`. Several of our dependencies do not support odd numbered releases. Rather than be restrictive, only including the affected versions for now.

Made some slight modifications to the `README` and `CONTRIBUTING` to plug a recommendation to use LTS versions of Node.

Closes #7816