feat(custom-resources): state machine provider
Add a state machine custom resource provider.

It uses the nested workflow pattern where the "user" state machine is started
as a task in the provider state machine. This allows to set a timeout for the
custom resource.

This could be used as a base for a Fargate task custom resource provider.

Closes #6505