fix(events): cannot use the same target account for 2 cross-account event sources
We hard code the SID of the EventBusPolicy that we generate in the account
of the target of a cross-account CloudWatch Event rule.
Which means that, if you have two sources in different accounts generating
events into the same target account,
you will get an error on CloudFormation deployment time about a duplicate SID.
Include the source account ID when generating the SID to make it unique.

Fixes #8010