chore: make bootstrap v2 completely qualified
Add `--qualifier` parameter to the CLI, and add the qualifier to all
role names. This makes it possible to have multiple copies of the
bootstrap v2 stack in the same account and region.

Having multiple copies of the bootstrap template in the same
environment is required (for example) for creating and using it in 
integration tests, which may run in an environment where a bootstrap
v2 stack already exists, or which may be run in parallel to each other
(and must not interfere).

Doing this required removing the `-bootstrap-` string from 
buckets and roles, they are now prefixed with `cdk-<qualifier>`
instead of `cdk-bootstrap-<qualifier>`, to stay within length restrictions
of IAM Role names. While we're at it, rename `deploy-action-role'
to just 'deploy-role' to focus less on the pipeline, and more accurately
reflect we're going to use the same roles to deploy using the CLI.

This is technically a breaking change, but we're not advertising or
recommending the new bootstrapping experience yet, so we should
get these last kinks out as soon as possible.

Also in this change (and the change motivator), re-enabling the
integration tests for the new bootstrapping and deployment experience.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*