chore: fix regression tests
The regression tests hotpatch the integ tests of (upcoming) 1.40.0 over
the released integ tests of 1.39.0, to work around a bug that existed in
the tests of 1.39.0 which prevents it from running properly.

In #7910 we introduced a new file, `jest.setup.js`, which is
referenced by the test runner that we copy over into the 1.39.0 branch.
Since that hotpatching script wasn't aware of that new file, it's not
copying it over, which is making the test fail.