chore(core): already-uploaded template URL is incorrect in manifest
(This is a "chore" instead of a "fix" because the feature hasn't been
released yet, and so there's no point drawing attention to it in the
CHANGELOG).

Turns out I missed a test while writing this feature and therefore a
certain aspect of itdidn't work (uncovered by the integ test that I
optimistically wrote against it).

The gist of it is this:

* For new-style synthesis, the stack's template is added to the
  asset manifest, as Yet Another Asset, and it will be uploaded
  by `cdk-assets`.
* When calling CloudFormation, the CLI needs to pass the `https://`
  URL to the template file (S3 REST API URL). In order to get that
  URL, the framework writes the S3 URL that it gets back from
  `addFileAsset()` to the manifest.
* Surprise! That was actually an invalid operation!
  `assetLocation.s3Url` is a Token containing CloudFormation
  expressions, and can't be written to the manifest in that way!
  In fact, in the manifest we get `"${Token.10}"`.
* Surpise 2! We can't actually write the correct URL in cleartext,
  because it depends on `${AWS::URLSuffix}` for which we don't
  have the actual values in the framework. Ever.

Solve this by writing a placeholder URL to the manifest (we
write `s3://.../...` instead). The CLI will convert the `s3://`
format to an `https://` format at the last moment, just before
calling CloudFormation.

In addition, manually bumping the cloud assembly version to accommodate for the fact that while the structure of the schema hasn't changed, the meaning of one of its properties has, resulting in a change to the CLI that should be force users to upgrade.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*