feat(core): introduce CfnJson to enable intrinsics in hash keys
Rarely there is a need to use a CloudFormation intrinsic in a JSON hash key. This is impossible since intrinsics are not strings. To circumvent that, CfnJson can be used to "delay" the rendition of the object using a simple custom resource
which simply parses and reflects the input value as an attribute.

Required in order to implement EKS IRSA (#8084) and other federated identity use cases.