feat(stepfunctions): new service integration classes for Lambda, SNS, and SQS (#7946)
merge functionality currently modeled under task and state into a single object that represents a task state as an abstract class. Service integrations extend the base class and add service specific configuration, metrics, and policies.

this commit introduces the "new" service integrations for Lambda, SNS, and SQS

Motivation: 
The current service integrations that are offered in `aws-stepfunctions-tasks` all currently implement a `bind()` method and contribute a portion of the configuration that creates a `Task` state.

However, it's often useful to configure state level properties such as paths, retries, errors based on the service integration and the pattern that has been requested.

Implementation:
* Duplicate the current `Task` class and merge the properties of a task state and a task service integration into a new abstract base class.
* Concrete implementation per service integration class where all of the best practices and user intents can be encoded
* After all the service integrations have been migrated, we will want to retire the `Task` class as well since we don't want it to be instantiated.

Paves the way for:
#6715 by simplifying the invocation of service integration calls. We would be able to start adding properties such as retries and errors.

#6489 where we will be making service integrations consistent

### End Commit Message

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*