fix(autoscaling): add noDevice as a volume type (#7253)
## Commit Message
fix(autoscaling): setting mappingEnabled to true results in a deployment error 

fixes #7242

Currently it's impossible to disable a mapping because the volume property is
required on a BlockDevice object. CloudFormation will fail at deploy time
with a validation failure because NoDevice can not be present at the same time
as Ebs or VirtualName.

## End Commit Message

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_