feat(stepfunctions-tasks): task for submitting a job to AWS batch
replacement for the current implementation of `RunBatchJob` where
service integration and state level properties are merged.

Follows the new integration pattern. Notable differences from the
`RunBatchJob` implementation:

* `payload` prop is now of type `sfn.TaskInput`
Rationale: old implementation precluded using task input as the
payload directly. Added a test for this as well.

* `timeout` in the old implementation has been renamed to `attemptDuration`
Rationale: old implementation was really configuring the sub-property
named `AttemptDurationSeconds` and precluded the ability to set a timeout
on the task itself.

Updated the README.
Note that the other unit tests and integ test have been left verbatim. This
is a light sanity test that expected templates have not changed.