chore(cli): simpler error message on cloud assembly version mismatch (#8129)
The error message we currently print when a user has a CLI version that is not compatible with the framework is:

```console
Cloud assembly schema version mismatch: Maximum schema version supported is 1.33.0, 
but found 2.0.0. Please upgrade your CLI in order to interact with this app.
```

This is pretty cryptic, we shouldn't be starting the message with `Cloud Assembly`, as most users aren't really versed in, or even aware of it. In addition, the versions mentioned here are confusing and might give the impression we are asking users to upgrade to CLI version `2.0.0`. 

This PR simplifies the message to:

```
❯ cdk synth                                                                                                                                                                                                                                                                                     [14:27:15]
The CLI version you are using does not support your application version. Please upgrade the CLI to the latest version.
(Your application requires a CLI that supports a cloud assembly of version '2.0.0' or above.)
```

It clearly states what wrongs (CLI version incompatible) and how to fix (upgrade) right of the bat. 
The tail of the message is an attempt to explain why the CLI does not support the framework. 

Fixes #7901

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*