feat(cfn-include): add support for Conditions
This change adds the capability to retrieve CfnCondition objects
from the template ingested by the CfnInclude class,
using a new method getCondition.
It also correctly populates the cfnOptions.condition
property of the L1 resources if they use the Condition resource attribute,
as well as adding support for the Fn::Equals function.