feat(eks): improve security using IRSA conditions (#8084)
Currently the `ServiceAccount`construct creates a role with no conditions to the trust relationship or assume role policy. Without this it is possible for other pods in the same namespace to assume the role. To tighten this security the conditions needs to be set.

Documentation: https://docs.aws.amazon.com/eks/latest/userguide/create-service-account-iam-policy-and-role.html#create-service-account-iam-role

- [x] Add condition to the policy document using a custom resource
- [x] Add unit tests
- [x] Add integration tests
- [x] Adjust README and remove warning

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*