feat(s3): supports RemovalPolicy for BucketPolicy
Exposes RemovalPolicy on the BucketPolicyProps so bucket policies can be
retained when a stack is deleted.

I'm conflicted about this implementation, because it seems the more recommended
way to create/associate policies with buckets is to use the addToResourcePolicy
(or the grant*) methods. One option would be to have the addToResourcePolicy
call set the removal policy on the bucket policy to whatever the policy is for
the bucket itself; however, that would be a backwards-incompatible change, and
so would need a feature flag.

Curious for feedback from the core members on this approach.

fixes #7415