feat(cli): new deployment monitoring
Implement a new stack monitor, which updates the screen in-place and
shows the resources currently being deployed plus a progress bar,
instead of dumping a giant history of all events to the screen.

The new monitor can be activated by setting `$CDK_FANCY_MONITOR` to a
truthy value. We might consider making it the default (I do think it's
better) but it might not work properly on Windows and on narrow screens.

Fix in both old and new stack monitors: detect rollback events, and
instead of increasing the event counter (leading to a progress of
`84/45` resources done), roll the counter backwards as resources are
being rolled back.