feat(cli): MFA support
With this change AWS CDK supports MFA. Specifically it
will support mfa_serial field in the profile config by asking
user for MFA token for the mfa_serial field ARN.

AWS SDK has support for this built in so only change is adding
tokenCodeFn function to sharedIniCredentials options. Callback
sent to that function is used to return token back to SDK.

Inquirer package is used to create interactive prompt for user
to type the MFA token.