chore: fix tsc-3.9 issues ahead of upgrade (#8182)
Fixes one issue where the typechecker in 3.9 is stricter when matching
type intersections. The particular issue was with a `string`-valued
`enum` attempting to match against the `string` type. Added a better
typed guard for this particular case fixed it.

Additionally, the new incremental build support would cause certain
`.json` files to not be `require`-able due to not being listed under
`include` in the `tsconfig.json` file generated by `jsii`. Instead of
copying the SDK metadata JSON document from the `aws-sdk` package,
inlined the data in a `.generated.ts` module, which provides a cleaner
type structure.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*