chore: add a base VSCode config
VSCode configs for our repo are tricky enough that people would benefit
from having them checked into the repo. Some people are strongly opposed
to having them checked in at the default location though, for what I
assume are the following reasons:

- There's no good way to have user-specific, workspace-specific
  preferences, so one set of `.vscode` files would apply to everyone.
- If you already had workspace-specific VSCode preferences, the new
  files would collide.
- Not everyone uses VSCode, so if we start adding `.vscode` files,
  we should also start adding `.idea` files and others, and where will
  it end?

As a compromise, adding a script which will copy a base VSCode config
into place.  You can choose the run the script if you want it, and you
can choose not to run it if you don't. Everybody happy, right?

If necessary, we'll be able to extend this in the future with custom
per-user configs, but for now let's start with something simple.