chore(core): add `Stage` class
Stages synthesize to embedded Cloud Assemblies.

Ideally there should be CLI support for embedded Cloud Assemblies, but
there isn't yet. It will be coming soon, I promise.

To implement this, had to grab the reins of synthesis back from
the `constructs` library. I've taken the liberty to declare
synthesis in `constructs` deprecated.

Took the opportunity where we're declaring a new type of construct
in the construct hierarchy to clean up the default stack name
generation (get rid of the hashes and add a prefix, which will
make deploying multiple Stages in the same environment
just naturally work out nicely in a non-ugly way).

Started to write a section in the README on Stacks and
Stages, but I couldn't figure out how to write about the
use for Stages without referencing unreleased things, so
I gave up.

Also moved the "artifacts" classes in `cx-api` together to
unclutter the file structure.